Extract lib/playlistified_entry.rb
