+++ app/models/entry.rb
@@ -1,40 +1,7 @@
 # coding: utf-8
 require('pink_spider')
 require('paginated_array')
-
-class PlaylistifiedEntry
-  attr_reader( :id, :url, :title, :descriptions,
-               :visual_url, :locale, :tracks, :entry)
-  def initialize(id, url, title, description, visual_url, locale, tracks, entry)
-    @id          = id
-    @url         = url
-    @title       = title
-    @description = description
-    @visual_url  = visual_url
-    @locale      = locale
-    @tracks      = tracks
-    @entry       = entry
-  end
-
-  def create_tracks
-    tracks = @tracks.map do |t|
-      track = Track.find_or_create_by(id: t['id'],
-                                      provider: t['provider'],
-                                      identifier: t['identifier']) do
-        puts "New track #{t['provider']} #{t['identifier']}}"
-      end
-      track.url = Track::url t['provider'], t['identifier']
-      EntryEnclosure.find_or_create_by entry:          @entry,
-                                       enclosure:      track,
-                                       enclosure_type: Track.name do
-        puts "Add new track #{t['provider']} #{t['identifier']} " +
-             "to entry #{@entry.id}"
-      end
-      track
-    end
-    tracks
-  end
-end
+require('playlistified_entry')
 
 class Entry < ApplicationRecord
   belongs_to :feed            , touch: true
@@ -218,12 +185,12 @@ class Entry < ApplicationRecord
     hash = PinkSpider.new.playlistify url: url, force: force
     return if hash.nil?
     PlaylistifiedEntry.new(hash['id'],
-                 hash['url'],
-                 hash['title'],
-                 hash['description'],
-                 hash['visual_url'],
-                 hash['locale'],
-                 hash['tracks'], self)
+                           hash['url'],
+                           hash['title'],
+                           hash['description'],
+                           hash['visual_url'],
+                           hash['locale'],
+                           hash['tracks'], self)
   end
 
   def as_json(options = {})
